added wifi probabilistic error

+++ saviSimulation/savi/jason_processing/UgVBehavior.java
@@ -34,7 +34,7 @@ public class UgVBehavior extends UxVBehavior {
 	 * @param type
 	 * @param initialPosition
 	 */
-	public UgVBehavior(String id, String type, PVector initialPosition, double reasoningCyclePeriod,  double sensorsErrorProb, double sensorsErrorStdDev, double seed) {	
+	public UgVBehavior(String id, String type, PVector initialPosition, double reasoningCyclePeriod,  double sensorsErrorProb, double sensorsErrorStdDev, int seed) {	
 		// Initialize data values
 		super(id, type, initialPosition,reasoningCyclePeriod, sensorsErrorProb, sensorsErrorStdDev, seed);
 	}
